feat: 딜러 구현

+++ src/main/kotlin/domain/player/state/Bust.kt
@@ -1,6 +1,6 @@
 package domain.player.state
 
 class Bust(other: PlayerState) : Finished(other) {
-    override fun earningRate(): Double = (-1).toDouble()
+    override fun earningRate(): Double = -1.0
     override fun score(): Int = Int.MIN_VALUE
 }
