+++ src/main/kotlin/domain/player/state/Stay.kt
@@ -1,6 +1,6 @@
 package domain.player.state
 
 class Stay(other: PlayerState) : Finished(other) {
-    override fun earningRate(): Double = (1).toDouble()
+    override fun earningRate(): Double = 1.0
     override fun score(): Int = cards.score()
 }
